Fix Annotation bug

+++ dessert_task/src/main/java/com/shen/dessert_task/annotation/HelperAnnotation.kt
@@ -13,9 +13,9 @@ annotation class Executors {
 }
 
 @IntDef(value = [
-    Priorities.THREAD_PRIORITY_FOREGROUND.toLong(),
-    Priorities.THREAD_PRIORITY_BACKGROUND.toLong(),
-    Priorities.THREAD_PRIORITY_LOWEST.toLong()
+    Priorities.THREAD_PRIORITY_FOREGROUND,
+    Priorities.THREAD_PRIORITY_BACKGROUND,
+    Priorities.THREAD_PRIORITY_LOWEST
 ])
 @Retention(AnnotationRetention.SOURCE)
 annotation class Priorities {
